Store the size of H3HashArray in a uint32_t instead of int8_t to avoid an overflow of the size.

+++ source/WUD/content/partitions/volumes/H3HashArray.h
@@ -28,5 +28,5 @@ public:
     ~H3HashArray();
 
     std::unique_ptr<uint8_t[]> data;
-    uint8_t size;
+    uint32_t size;
 };
